ChatService_FastApi/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv

sqlalchemy
httpx
psycopg2-binary
